Changes to Bitmap64 after build
