--- extensions-contrib/druid-exactcount/src/main/java/org/apache/druid/query/aggregation/exactcount/bitmap64/Bitmap64ExactCountPostAggregator.java
@@ -1,138 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.druid.query.aggregation.exactcount.bitmap64;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonTypeName;
-import com.google.common.collect.Sets;
-import org.apache.druid.query.aggregation.AggregatorFactory;
-import org.apache.druid.query.aggregation.PostAggregator;
-import org.apache.druid.query.aggregation.post.ArithmeticPostAggregator;
-import org.apache.druid.query.aggregation.post.PostAggregatorIds;
-import org.apache.druid.query.cache.CacheKeyBuilder;
-import org.apache.druid.segment.ColumnInspector;
-import org.apache.druid.segment.column.ColumnType;
-
-import javax.annotation.Nullable;
-import java.util.Comparator;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Set;
-
-@JsonTypeName("bitmap64ExactCountCardinality")
-public class Bitmap64ExactCountPostAggregator implements PostAggregator
-{
-  private final String name;
-  private final String fieldName;
-
-  @JsonCreator
-  public Bitmap64ExactCountPostAggregator(
-      @JsonProperty("name") String name,
-      @JsonProperty("fieldName") String fieldName
-  )
-  {
-    this.name = name;
-    this.fieldName = fieldName;
-  }
-
-  @JsonProperty
-  public String getFieldName()
-  {
-    return fieldName;
-  }
-
-  @Override
-  public Set<String> getDependentFields()
-  {
-    return Sets.newHashSet(fieldName);
-  }
-
-  @Override
-  public Comparator getComparator()
-  {
-    return ArithmeticPostAggregator.DEFAULT_COMPARATOR;
-  }
-
-  @Override
-  @JsonProperty
-  public Object compute(Map<String, Object> combinedAggregators)
-  {
-    Object value = combinedAggregators.get(fieldName);
-    return ((Bitmap64Counter) value).getCardinality();
-  }
-
-  @Override
-  @JsonProperty
-  public String getName()
-  {
-    return name;
-  }
-
-  @Nullable
-  @Override
-  public ColumnType getType(ColumnInspector signature)
-  {
-    return ColumnType.LONG;
-  }
-
-  @Override
-  public PostAggregator decorate(Map<String, AggregatorFactory> aggregators)
-  {
-    return this;
-  }
-
-  @Override
-  public byte[] getCacheKey()
-  {
-    return new CacheKeyBuilder(PostAggregatorIds.BITMAP64_EXACT_COUNT_MERGE_TYPE_ID)
-        .appendString(fieldName)
-        .build();
-  }
-
-  @Override
-  public String toString()
-  {
-    return "Bitmap64ExactCountPostAggregator{" +
-           "name='" + name + '\'' +
-           ", field=" + fieldName +
-           '}';
-  }
-
-  @Override
-  public boolean equals(Object o)
-  {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    Bitmap64ExactCountPostAggregator that = (Bitmap64ExactCountPostAggregator) o;
-    return name.equals(that.name) &&
-           fieldName.equals(that.fieldName);
-  }
-
-  @Override
-  public int hashCode()
-  {
-    return Objects.hash(name, fieldName);
-  }
-}
