+++ pom.xml
@@ -233,7 +233,7 @@
         <module>extensions-contrib/cloudfiles-extensions</module>
         <module>extensions-contrib/graphite-emitter</module>
         <module>extensions-contrib/distinctcount</module>
-        <module>extensions-contrib/druid-exactcount</module>
+        <module>extensions-contrib/druid-exact-cardinality</module>
         <module>extensions-contrib/statsd-emitter</module>
         <module>extensions-contrib/time-min-max</module>
         <module>extensions-contrib/virtual-columns</module>
